+++ coaster/logger.py
@@ -17,7 +17,7 @@ import requests
 from pprint import pprint
 import six
 from flask import g, request, session
-from .user import current_user
+from .auth import current_auth
 
 
 # global var as lazy in-memory cache
@@ -111,8 +111,8 @@ class LocalVarFormatter(logging.Formatter):
             pprint_with_indent(vars(g), sio)
 
         print('\n----------\n', file=sio)
-        print("Current user:", file=sio)
-        pprint_with_indent(vars(current_user), sio)
+        print("Current auth:", file=sio)
+        pprint_with_indent(vars(current_auth), sio)
 
         s = sio.getvalue()
         sio.close()